requirements.txt
aiofiles==23.2.1
altair==5.3.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.4.0
asttokens==2.4.1
attrs==23.2.0
blinker==1.8.2
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
comm==0.2.2
conda-pack==0.8.1
contourpy==1.2.1
cycler==0.12.1
debugpy==1.6.7
decorator==5.1.1
dnspython==2.6.1
email_validator==2.2.0
entrypoints==0.4
exceptiongroup==1.2.1
executing==2.0.1
fastapi==0.115.6
fastapi-cli==0.0.7
ffmpy==0.3.2
filelock==3.15.4
Flask==3.0.3
fonttools==4.53.1
fsspec==2024.6.1
gradio==4.44.1
gradio_client==1.3.0
gradio_image_prompter==0.1.0
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.27.0
idna==3.7
imageio==2.35.1
importlib_resources==6.4.0
ipykernel==6.29.5
ipython==8.26.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
jupyter-client==7.3.4
jupyter_core==5.7.2
kiwisolver==1.4.5
lazy_loader==0.4
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.1
matplotlib-inline==0.1.7
mdurl==0.1.2
nest_asyncio==1.6.0
networkx==3.3
numpy==2.0.0
omegaconf==2.3.0
opencv-contrib-python==4.10.0.84
opencv-python==4.10.0.84
orjson==3.10.6
packaging==24.1
pandas==2.2.2
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.4.0
pip==24.0
platformdirs==4.2.2
prompt_toolkit==3.0.47
psutil==5.9.0
ptyprocess==0.7.0
pure_eval==0.2.3
pydantic==2.10.4
pydantic_core==2.27.2
pydantic-extra-types==2.10.1
pydantic-settings==2.7.0
pydub==0.25.1
Pygments==2.18.0
pyparsing==3.1.2
python-dateutil==2.9.0
python-dotenv==1.0.1
python-multipart==0.0.20
pytz==2024.1
PyYAML==6.0.1
pyzmq==25.1.2
rawpy==0.23.1
referencing==0.35.1
requests==2.32.3
rich==13.7.1
rich-toolkit==0.12.0
rpds-py==0.18.1
ruff==0.5.1
safehttpx==0.1.6
scikit-image==0.24.0
scipy==1.14.1
semantic-version==2.10.0
setuptools==69.5.1
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
stack-data==0.6.2
starlette==0.41.3
tifffile==2024.8.28
tomlkit==0.12.0
toolz==0.12.1
tornado==6.1
tqdm==4.66.4
traitlets==5.14.3
typer==0.12.3
typing_extensions==4.12.2
tzdata==2024.1
ujson==5.10.0
urllib3==2.2.2
uvicorn==0.30.1
uvloop==0.19.0
watchdog==5.0.2
watchfiles==0.22.0
wcwidth==0.2.13
websockets==11.0.3
Werkzeug==3.0.4
wheel==0.43.0
